ml_service/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0

# Machine Learning
torch==2.2.0
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0

# Optional: Deep Learning extras
# transformers==4.37.0
# sentence-transformers==2.3.1
